test: Add more tests for the VersionParser (#55)
Using the [Semver
specification](https://semver.org/#is-there-a-suggested-regular-expression-regex-to-check-a-semver-string),
I added more tests.

I preserved a few tests on top of those ones and I added the important,
potentially unorthodox cases specific to the PhpSpec version.

Despite a few cases that seems buggy, the ones about the PhpSpec version
are all covered except `dev-main`, which I do not think is too big of a
deal (at least it doesn't need to be addressed urgently).