ci: Add a workflow to automatically update the infection PHAR version used (#53)
This workflow triggers via a cron job & can be manually triggered.

I don't think it's too great of a system, ideally we should have
continuous testing not just when a new release comes out. But this would
require more work and I'm not too adept on GitHub Actions either. So
this PR is more of a compromise.