fix: Update the regex to better detect more cases (#60)
From the spec I could find a regex so I wanted to apply it. It still
required some adaptions though.

I think ultimately it is fine for it to not be completely correct as
what is important is that the version is usable for `version_compare()`,
but if it is still a problem I think ultimately we may need multiple
regexes.