perf: Do not fetch the content of the original configuration more than once (#68)
We've been using `Yaml::parseFile` in `InitialConfigBuilder` and
`MutationConfigBuilder`.

I think one reason for it, is that the PhpSpec configuration builder
itself is mutable, so we get a fresh instance in each and there is no
risk of corrupting the data of another mutation. However, in its current
form, this means we not only do the parsing of the YAML everytime, but
also fetching the file from the filesystem, which is easily preventable.

In this PR, we fetch the content of the PhpSpec configuration file (not
decoded) when building the adapter instance. As a result, we can forward
that data, reducing the filesystem calls from `initial + nbrOfMutation`
to 1.