fix: Require at least one passing test for the test output to be considered as valid (#87)
In theory it is a big change because when previously a test suite was
completely skipped or no test is executed, it would pass and won't
anymore.

In practice however, this is unlikely to cause any big change as if
there wasn't any test in the first place, the code wouldn't be
considered covered by a test framework hence not mutated.