refactor: Inject the adapter name via the factory (#89)
Having `PhpSpecAdapter` provide its name makes total sense. However, I
noticed for example `VersionParser` that also needs access to that name
somehow to provide a better exception message to the user. Currently it
is hardcoded, and it is not a problem.

However, ultimately, I would like `VersionParser` to be re-usable and
come from somewhere else, i.e. not implemented from this package. To
this end, it is better of the test framework name is injected instead.
Since those services are instantiated from the Factory, and the adapter
is not meant to be created without its factory anyway, I think it's fine
to have the name declared by the adapter factory and injected instead.
As a result, we can inject the name to the instantiated services too,
regardless of which package they were declared.