refactor: Inject the coverage directory from the factory (#93)
What prompted this is I was curious how infection found the `index.xml`
at all in the first place.

Turns out we do not rely at all on the expected `index.xml` path, but we
still find the file because we use the finder to lookup in the parent
directory.

Using the default path is better because more performant. But knowing it
can only be done by `infection`. So to forward this information to the
adapters, we need to adapt the factory.

I putted this idea in
https://github.com/infection/infection/issues/2760, meanwhile this PR
only contents itself with letting the factory define the coverage path
used. I will update the coverage path to match with the default one in a
separate PR (as it has behaviour changes, it would no longer is a
refactoring).

A nice side-effect is that it showcase when we're really using the
injected `tmpDirectory` and when we are using the
`coverageDirectoryPath`.